specify int64 to prevent different behaviour on different platforms

+++ onnx/backend/test/case/node/mod.py
@@ -194,9 +194,9 @@ class Mod(Base):
             outputs=['z'],
         )
 
-        x = np.arange(0, 30).reshape([3, 2, 5])
-        y = np.array([7])
-        z = np.mod(x, y)
+        x = np.arange(0, 30).reshape([3, 2, 5]).astype(np.int64)
+        y = np.array([7]).astype(np.int64)
+        z = np.mod(x, y).astype(np.int64)
         z
         #   array([[[0, 1, 2, 3, 4],
         #     [5, 6, 0, 1, 2]],
